Content Warning (Main post only)
Images with any kind of content warning label whatsoever from the Bluesky side will no longer display. Instead, a content warning will be shown, displaying the specific flag types, and a link to visit the actual post on Bluesky instead.

Unfortunately, we've discovered that none of the links to the Bluesky post post seem to actually work.

Content Warnings only show for posts/skeets and non-quoted reposts/reskeets with images that are given a content warning. We'll need test data from reverendcrush to proceed further. Until then, next step is to fix the links to the post.

SCSS files updated for new feature.