ENDGAME! Improved Banger Score WITH LEADERBOARD!!
We have implemented a new version of the Banger Score. In addition to adjusted scoring, there's also a "Ratio Penalty" for posts with 95% fewer likes than replies. We have NO idea how well this is going turn out, but we'll just have to test it once it goes live and kind of study the data.

We also implemented a leaderboard of the 25 posts with the highest Banger Scores. It's rather barebones, looking only from the pool of the previous 100 posts, but for demonstration purposes, it does the job!

All that's really left is some styling, and THIS portion of the project, which is 90% of the work, will be ready for v1.0!