Leaderboard styles
The bulk of Leaderboard design work has been set; it looks good on mobile, but will have to be adapted for tablet moving into lap/desktop sizes (yes, we're taking tablet into consideration since folks like to resize browsers anyway.)

Also, for the actual main site... the canonical body text for the app has been changed from Righteous to Jost. This will probably change when we get to developing the rest of the site, as Jost was a holdover from the old MCT630 project. ReverendCrush is going to want something less "Pretentious Web 2.0 Horses**ty" as he puts it... but we'll cross that bridge when we get to it.