Early content, no styling
This is still early, but there's not going to be a whole lot going on with these pages (HomeComponent.js is more or less ready.) It's really just adding some content and styling it. This is just laying down the ground converting this over from it's old MCT6|30 roots. Still waiting on Reverendcrush for info.