V.1.0.0! (Soft Launch)
This should be everything, hopefully. Launch day is 3/19, tomorrow, if everything goes according to plan.

Because I expect this Git repository to be shared tomorrow, I replaced the App Password to a placeholder, though Rev is going to whip up a new one for build either later today or very early tomorrow.