QRS Image Fix & Shorter CTA
Some images from quote posts of reposted posts weren't showing up. That's been fixed.

Also, more noticable, is shortening the call to action to follow reverendcrush on Bluesky. It was a mouthful that wasn't needed, and got in the way of the navigation. It's been moved to the very bottom of the component and greatly shortened.